src/backend/email_sender/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==1.0.0 